Adding indent rules when inserting parentheses
When a new function is created and you type the closing bracket, you
would always have to run the indent. For me it would never work and
would always look like this:

```
test :: () -> s32 {
    }
```

This new indent hook will run the indent rules on inserting special
characters. The function itself was taken from a different major mode
and used here, however I feel like when the language matures the entire
indenting rules should be migrated to be custom for Jai.